aula_2/projects/email-vagas-ia/pyproject.toml
[project]
name = "email-vagas-ia"
version = "0.1.0"
description = ""
authors = [
    {name = "EonCyber",email = "eonciber@gmail.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "langchain-core (>=0.3.65,<0.4.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain-google-community (>=2.0.7,<3.0.0)",
    "keyflow (>=1.1.1,<2.0.0)",
    "keyboard (>=0.13.5,<0.14.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "langchain-docling (>=1.0.0,<2.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
]

[tool.poetry]
packages = [
    {include = "mail", from = "src"},
    {include = "rag", from = "src"},
    {include = "watcher", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
